REEEEEEEEEEE

+++ cogs/mod.py
@@ -88,9 +88,12 @@ class mod:
         if not ctx.guild.me.guild_permissions.kick_members:
             return await ctx.send("Hey! How do you expect me to handle raidmode without the **Kick Members** permission?\n\nPlease grant me this permission, and re-run this command.")
         if action.lower() == "on":
-            await self.bot.db.raidmode.update_one({"id": ctx.guild.id}, {"$set": {"status": True}})
+            await self.bot.db.raidmode.update_one({"id": ctx.guild.id}, {"$set": {"status": True}}, upsert=True)
             return await ctx.send("Alright, raidmode is **on.** I'll kick any bakas joining this server after this. ***LET 'EM AT ME!***")
         elif action.lower() == "off":
+            data = await self.bot.db.raidmode.find_one({"id": ctx.guild.id})
+            if not data:
+                return await ctx.send("Raidmode was never on, you baka. What are you trying to get from me??")
             await self.bot.db.raidmode.delete_one({"id": ctx.guild.id})
             return await ctx.send("Alright, raidmode is **off.** Letting my guard down now...")
         else:
